+++ package.json
@@ -22,6 +22,7 @@
     "styled-components": "^6.1.19"
   },
   "devDependencies": {
+    "@types/mapbox-gl": "^3.4.1",
     "babel-plugin-styled-components": "^2.1.4",
     "eslint": "^8",
     "eslint-config-next": "14.2.4",
