+++ src/components/MtaMap/mtaMapOptions.ts
@@ -1,10 +1,11 @@
 import mapStyle from "@/styles/mapbox-style.json"; 
+import { StyleSpecification } from "mapbox-gl";
 
 export const getMtaMapOptions = (container) => {
    return ({
     container: container,
-    style: mapStyle,
-    center: [-73.98365318925187, 40.7583063693059], // NYC
+    style: mapStyle as unknown as StyleSpecification,
+    center: [-73.98365318925187, 40.7583063693059] as [number, number], // NYC
     zoom: 13,
    })
 }
