+++ src/components/icons/index.js
@@ -1,4 +1,4 @@
 export { default as AccessibleIcon } from './AccessibleIcon';
 export { default as ElevatorOutIcon} from './ElevatorOutIcon';
 export { default as WarnIcon} from './WarnIcon'
-export {default as AccessibleIconWhite} from './AccessibleIconWB'
+export {default as AccessibleIconWhite} from './AccessibleIconWhite'
